backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
aiogram>=3.13.0
httpx>=0.27.2
python-multipart>=0.0.12
python-dotenv>=1.0.0
